+++ includes/integrations/class.EwayPaymentsEventsManager.php
@@ -320,17 +320,16 @@ class EwayPaymentsEventsManager extends EM_Gateway {
 		// maybe set up Client Side Encryption
 		$creds = $this->getApiCredentials();
 		if (!empty($creds['ecrypt_key'])) {
-			wp_enqueue_script('eway-ecrypt');
-			add_action('wp_print_footer_scripts', array($this, 'ecryptScript'));
+			wp_enqueue_script('eway-payment-gateway-ecrypt');
+			add_action('wp_footer', array($this, 'ecryptScript'));
 		}
 	}
 
 	/**
-	* inline scripts for client-side encryption
+	* configure the scripts for client-side encryption
 	*/
 	public function ecryptScript() {
 		$creds	= $this->getApiCredentials();
-		$min	= SCRIPT_DEBUG ? '' : '.min';
 
 		$vars = array(
 			'mode'		=> 'events-manager',
@@ -342,10 +341,7 @@ class EwayPaymentsEventsManager extends EM_Gateway {
 						),
 		);
 
-		echo '<script>';
-		echo 'var eway_ecrypt_vars = ', json_encode($vars), '; ';
-		readfile(EWAY_PAYMENTS_PLUGIN_ROOT . "js/ecrypt$min.js");
-		echo '</script>';
+		wp_localize_script('eway-payment-gateway-ecrypt', 'eway_ecrypt_vars', $vars);
 	}
 
 	/**